+++ Assets/Scripts/BigBoss2.cs
@@ -40,7 +40,7 @@ public class BigBoss2 : Mob, IMapObject
 
 	private Char[] charAttack;
 
-	private int[] dameHP;
+	private long[] dameHP;
 
 	private sbyte type;
 
@@ -301,7 +301,7 @@ public class BigBoss2 : Mob, IMapObject
 		}
 	}
 
-	public void setAttack(Char[] cAttack, int[] dame, sbyte type)
+	public void setAttack(Char[] cAttack, long[] dame, sbyte type)
 	{
 		status = 3;
 		charAttack = cAttack;
@@ -327,7 +327,7 @@ public class BigBoss2 : Mob, IMapObject
 			{
 				for (int i = 0; i < charAttack.Length; i++)
 				{
-					charAttack[i].doInjure(dameHP[i], 0.0, isCrit: false, isMob: false);
+					charAttack[i].doInjure(dameHP[i], 0, isCrit: false, isMob: false);
 					ServerEffect.addServerEffect(102, charAttack[i].cx, charAttack[i].cy, 1);
 				}
 			}
@@ -344,7 +344,7 @@ public class BigBoss2 : Mob, IMapObject
 			{
 				for (int j = 0; j < charAttack.Length; j++)
 				{
-					MonsterDart.addMonsterDart(x + ((dir != 1) ? (-45) : 45), y - 25, isBoss: true, dameHP[j], 0.0, charAttack[j], 24);
+					MonsterDart.addMonsterDart(x + ((dir != 1) ? (-45) : 45), y - 25, isBoss: true, dameHP[j], 0, charAttack[j], 24);
 				}
 			}
 		}
@@ -365,7 +365,7 @@ public class BigBoss2 : Mob, IMapObject
 		{
 			for (int k = 0; k < charAttack.Length; k++)
 			{
-				charAttack[k].doInjure(dameHP[k], 0.0, isCrit: false, isMob: false);
+				charAttack[k].doInjure(dameHP[k], 0, isCrit: false, isMob: false);
 				ServerEffect.addServerEffect(102, charAttack[k].cx, charAttack[k].cy, 1);
 			}
 		}
